Разбил компонент комментариев на две части
