+++ src/components/film-details.js
@@ -146,6 +146,7 @@ export default class FilmDetails extends AbstractSmartComponent {
 
   recoveryListeners() {
     this.setPopupCloseClickHandler(this._submitHandler);
+
     this._subscribeOnEvents();
   }
 
@@ -163,6 +164,10 @@ export default class FilmDetails extends AbstractSmartComponent {
 
     this._submitHandler = handler;
   }
+
+  // setCommentAddClickHandler(handler) {
+  //   this.getElement().querySelector(`form`).addEventListener(`keydown`, handler);
+  // }
 
   _subscribeOnEvents() {
     const element = this.getElement();
